requirements.txt
Flask
pandas
matplotlib
statsmodels
python-docx